More chmod +x fixes
It seems that switching to libzipsharp made permissions on all files just a bit worse.  There's probably a reasonable workaround however it did not solve the problem it was supposed to so there's no need to use it just now.

This switches back to system io compression and a couple more tweaks to the chmod calls on mac.